*** empty log message ***

+++ WEB-INF/config/database/mysql/mysql_db_struct.sql
@@ -336,7 +336,7 @@ CREATE TABLE jforum_users (
   deleted tinyint(1) default NULL,
   user_viewonline tinyint(1) default '1',
   security_hash varchar(32),
-  user_karma DECIMAL(10,2),
+  user_karma DOUBLE,
   user_authhash VARCHAR(32),
   PRIMARY KEY  (user_id)
 ) TYPE=InnoDB;
